+++ api/utils.py
@@ -9,7 +9,7 @@ import re
 class APIUtils():
     QUERY_TERMS = [
         'keyword', 'title', 'author', 'subject', 'viaf', 'lcnaf',
-        'date', 'startYear', 'endYear', 'language', 'format', 'showAll'
+        'date', 'startYear', 'endYear', 'language', 'format', 'govDoc', 'showAll'
     ]
 
     FORMAT_CROSSWALK = {
@@ -84,11 +84,11 @@ class APIUtils():
 
         for key, value in aggregations.items():
             if key == 'buckets':
+                print(key, value)
                 aggs[parentKey] = [
-                    {
-                        'value': b['key'],
-                        'count': b['editions_per']['doc_count']
-                    }
+                    {'value': b['key'], 'count': b['editions_per']['doc_count']}
+                    if 'editions_per' in b.keys() else 
+                    {'value': b['key_as_string'],'count': b['doc_count']}
                     for b in value
                 ]
                 return aggs