+++ tests/unit/test_api_es.py
@@ -1,5 +1,6 @@
 import pytest
 
+from elasticsearch_dsl import Search, Q, A
 from tests.helper import TestHelpers
 from api.elastic import ElasticClient, ElasticClientError
 
@@ -785,6 +786,47 @@ class TestElasticClient:
         assert testInstance.appliedFilters == ['displayFilter']
         assert testInstance.appliedAggregations == ['displayAggregation']
 
+    def test_createFilterClausesAndAggregations_w_noGovDoc(self, testInstance, mocker):
+        mockQuery = mocker.patch('api.elastic.Q')
+        mockQuery.side_effect = ['displayFilter', 'govDocFilters']
+        mockAgg = mocker.patch('api.elastic.A')
+        mockAgg.side_effect = ['displayAggregation', 'govDocAggregation']
+
+        testInstance.createFilterClausesAndAggregations([('govDoc', 'noGovDoc')])
+
+        mockQuery.assert_has_calls([
+            mocker.call('exists', field='editions.formats'),
+            mocker.call('term', **{'is_government_document': False})
+        ])
+        mockAgg.assert_has_calls([
+            mocker.call('filter', exists={'field': 'editions.formats'}),
+            mocker.call('filter', **{'term': {'is_government_document': False}})
+        ])
+
+        assert testInstance.govDocFilter == 'govDocFilters'
+        assert testInstance.appliedFilters == ['displayFilter']
+        assert testInstance.appliedAggregations == ['displayAggregation']
+
+    def test_createFilterClausesAndAggregations_w_onlyGovDoc(self, testInstance, mocker):
+        mockQuery = mocker.patch('api.elastic.Q')
+        mockQuery.side_effect = ['displayFilter', 'govDocFilters']
+        mockAgg = mocker.patch('api.elastic.A')
+        mockAgg.side_effect = ['displayAggregation', 'govDocAggregation']
+
+        testInstance.createFilterClausesAndAggregations([('govDoc', 'onlyGovDoc')])
+
+        mockQuery.assert_has_calls([
+            mocker.call('exists', field='editions.formats'),
+            mocker.call('term', **{'is_government_document': True})
+        ])
+        mockAgg.assert_has_calls([
+            mocker.call('filter', exists={'field': 'editions.formats'})
+        ])
+
+        assert testInstance.govDocFilter == 'govDocFilters'
+        assert testInstance.appliedFilters == ['displayFilter']
+        assert testInstance.appliedAggregations == ['displayAggregation']
+
     def test_createFilterClausesAndAggregations_no_filters(self, testInstance, mocker):
         mockQuery = mocker.patch('api.elastic.Q')
         mockAgg = mocker.patch('api.elastic.A')
@@ -836,6 +878,7 @@ class TestElasticClient:
         mockQuery.query.return_value = mockQuery
         testInstance.query = mockQuery
         testInstance.dateSort = None
+        testInstance.govDocFilter = None
         testInstance.appliedFilters = ['filter1', 'filter2']
         testInstance.languageFilters = []
 
@@ -856,6 +899,7 @@ class TestElasticClient:
         mockQuery.query.return_value = mockQuery
         testInstance.query = mockQuery
         testInstance.dateSort = None
+        testInstance.govDocFilter = None
         testInstance.appliedFilters = []
         testInstance.languageFilters = ['Lang1', 'Lang2']
 
@@ -869,6 +913,48 @@ class TestElasticClient:
         ])
         mockQuery.query.assert_called_once_with('bool', must=['Lang1Nested', 'Lang2Nested'])
 
+    def test_applyFilters_govDoc_filter_only(self, testInstance):
+
+        testInstance.query = Search().query(Q())
+        testInstance.dateSort = None
+        testInstance.languageFilters = []
+        testInstance.appliedFilters = []
+        testInstance.govDocFilter = Q('term', **{'is_government_document': True})
+
+        testInstance.applyFilters()
+
+        assert testInstance.query.to_dict()['query']['bool']['must'][0] == {'nested': { 
+                                            'inner_hits': {'size': 100},
+                                            'path': 'editions',
+                                            'query': {'match_all': {}}}}
+        assert testInstance.query.to_dict()['query']['bool']['must'][1] == {'term': {
+                                                'is_government_document': True}}
+                                          
+
+
+    def test_applyFilters_govDoc_and_other_filters(self, testInstance):
+
+        testInstance.query = Search().query(Q())
+        testInstance.dateSort = None
+        testInstance.languageFilters = []
+        testInstance.appliedFilters = [Q('term', editions__formats = 'testFormat')]
+        testInstance.govDocFilter = Q('term', **{'is_government_document': True})
+
+        testInstance.applyFilters()
+
+        assert testInstance.query.to_dict()['query']['bool']['must'][0] == {'nested': { 
+                                                'inner_hits': {'size': 100},
+                                                'path': 'editions',
+                                                'query': { 
+                                                'bool': { 
+                                                'must': [{
+                                                'term': { 
+                                                'editions.formats': 'testFormat'}}]
+                                                }}}}
+        assert testInstance.query.to_dict()['query']['bool']['must'][1] == {'term': {
+                                                'is_government_document': True}}
+        
+
     def test_applyFilters_language_and_other_filters(self, testInstance, mocker):
         mockClause = mocker.patch('api.elastic.Q')
         mockClause.side_effect = ['Lang1Filter', 'Lang1Nested']
@@ -877,6 +963,7 @@ class TestElasticClient:
         mockQuery.query.return_value = mockQuery
         testInstance.query = mockQuery
         testInstance.dateSort = None
+        testInstance.govDocFilter = None
         testInstance.appliedFilters = ['formatFilter']
         testInstance.languageFilters = ['Lang1']
 
@@ -896,6 +983,7 @@ class TestElasticClient:
         mockQuery.query.return_value = mockQuery
         testInstance.query = mockQuery
         testInstance.dateSort = {'editions.publication_date': {'test': 'value', 'nested': 'toDelete'}}
+        testInstance.govDocFilter = None
         testInstance.appliedFilters = []
         testInstance.languageFilters = []
 
@@ -909,28 +997,58 @@ class TestElasticClient:
             query='filterQuery'
         )
 
-    def test_applyAggregations(self, testInstance, mocker):
-        mockAgg = mocker.patch('api.elastic.A')
-        mockAgg.return_value = 'baseAgg'
-        mockRoot = mocker.MagicMock()
-        mockQuery = mocker.MagicMock()
-        mockQuery.aggs.bucket.return_value = mockRoot
-        mockRoot.bucket.return_value = mockRoot
+    def test_applyAggregation_govDoc_only(self, testInstance):
 
-        testInstance.query = mockQuery
-        testInstance.appliedAggregations = ['testAgg']
+        testInstance.query = Search()
+        testInstance.dateSort = None
+        testInstance.displayAggregation = None
+        testInstance.appliedAggregations = []
+        testInstance.govDocAgg = A('filter', **{'term': {'is_government_document': True}})
+
         testInstance.applyAggregations()
 
-        mockAgg.assert_called_once_with('nested', path='editions')
-        mockQuery.aggs.bucket.assert_called_once_with('editions', 'baseAgg')
-        mockRoot.bucket.assert_has_calls([
-            mocker.call('edition_filter_0', 'testAgg'),
-            mocker.call('lang_parent', 'nested', path='editions.languages'),
-            mocker.call('languages', 'terms', field='editions.languages.language', size=200),
-            mocker.call('editions_per', 'reverse_nested'),
-            mocker.call('formats', 'terms', field='editions.formats', size=10),
-            mocker.call('editions_per', 'reverse_nested')
-        ])
+        assert testInstance.query.aggs.to_dict()['aggs']['govDoc'] == {'terms': {'field': 'is_government_document', 'size': 2}}
+        assert testInstance.query.aggs.to_dict()['aggs']['govDoc_filter'] == {'filter': {'term': {'is_government_document': True}}}
+
+    def test_applyAggregations_govDoc_and_other_aggregations(self, testInstance):
+
+        testInstance.query = Search()
+        testInstance.dateSort = None
+        testInstance.displayAggregation = None
+        testInstance.appliedAggregations = [A(
+                'filter', **{'range': {'editions.publication_date': 'testDate'}}
+            ), A(
+                'filter', **{'terms': {'editions.formats': 'testFormat'}}
+            ), A('filter', **{'exists': {'field': 'editions.formats'}})]
+        testInstance.govDocAgg = A('filter', **{'term': {'is_government_document': False}})
+
+        testInstance.applyAggregations()
+        
+        assert testInstance.query.aggs.to_dict()['aggs']['govDoc'] == {'terms': {'field': 'is_government_document', 'size': 2}}
+        assert testInstance.query.aggs.to_dict()['aggs']['govDoc_filter'] == {'filter': {'term': {'is_government_document': False}}}
+        assert testInstance.query.aggs.to_dict()['aggs']['editions'] == {'aggs': 
+        {'edition_filter_0': 
+            {'aggs': 
+                {'edition_filter_1': 
+                    {'aggs': 
+                        {'edition_filter_2': 
+                            {'aggs': 
+                                {'formats': 
+                                    {'aggs': {'editions_per': {'reverse_nested': {}}},
+                                    'terms': {'field': 'editions.formats', 'size': 10}},
+                                    'lang_parent': 
+                                        {'aggs': 
+                                            {'languages': 
+                                                {'aggs': 
+                                                    {'editions_per': {'reverse_nested': {}}},
+                                                'terms': {'field': 'editions.languages.language', 'size': 200}}},
+                                            'nested': {'path': 'editions.languages'}}},
+                                        'filter': {'exists': {'field': 'editions.formats'}}}},
+                                    'filter': {'terms': {'editions.formats': 'testFormat'}}}},
+                                'filter': {'range': {'editions.publication_date': 'testDate'}}}},
+                            'nested': {'path': 'editions'}}
+                                                        
+
 
     def test_addSearchHighlighting(self, testInstance, mocker):
         mockQuery = mocker.MagicMock()